Multiple Windows, Thread Safety, and Bug Fixes, Oh My!
This commit adds the functionality to switch between different sets of
menus via buttons to the front end GUI. Also changes to the code now
make the GUI thread safe. Additionally, some bug fixes (mainly images
load in instantly now instead of after selecting a route) and cleaning
up much of the code to be more readable and less full of outdated
comments.