Disabling/Enabling Buttons, Icons, Loading Displays
The user is no longer able to press buttons they should not be able to.
Also adds an icon to replace the standard Java icon. Also when GUI or
MapUpdaterGUI is launched, the mouse cursor changes to a loading cursor
until the GUI is finished loading. Additionally, there is a loading gif
that should appear in the top right corner of MapUpdaterGUI (not
completely done yet). 

NOTE: These changes do not apply to the navMenu section of GUI. Still
waiting to implement changes with that.